+++ src/main/java/com/appcenter/marketplace/domain/tempMarket/repository/TempMarketRepositoryCustom.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface TempMarketRepositoryCustom {
     List<TempMarketRes> findMarketList(Long memberId, Long marketId, Integer size);
     List<TempMarketRes> findMarketListByCategory(Long memberId, Long marketId, Integer size, String category);
-    List<TempMarketRes> findUpcomingMarketList(Long memberId, Long marketId, Long cheerCount, Integer size);
+    List<TempMarketRes> findUpcomingMarketList(Long memberId, Long marketId, Integer cheerCount, Integer size);
 }
